feat(web): Jobs IS the cloud Dashboard
User: "不要jobs 就是放到首页 作为dashbroad". The /cloud/jobs page was
duplicating Dashboard — both rendered as nav entries, Dashboard just
forwarded to Jobs. Now:

  - Extract the Jobs & Runs UI into web/src/components/JobsBoard.tsx
    (was inline in _app.cloud.jobs.tsx).
  - _app.dashboard.tsx renders <JobsBoard /> in cloud mode (no more
    Navigate redirect) and the legacy LocalDashboard in local-web mode.
  - _app.cloud.jobs.tsx collapses to a Navigate /dashboard redirect so
    old bookmarks still resolve.
  - Top nav drops "Jobs" — Dashboard now IS the Jobs board on cloud.

Cloud nav becomes:
  Dashboard | Repos | Projects | Operators | Machines | Bindings | Settings
